Implement bucketed naive colocalizer

+++ src/main/kotlin/simplecolocalization/services/colocalizer/PositionedCell.kt
@@ -6,12 +6,12 @@ import ij.gui.Roi
  * The representation of a positioned cell is a set of points on a
  * two-dimensional coordinate system belonging which form the cell.
  */
-class PositionedCell(val points: Set<Pair<Float, Float>>) {
+class PositionedCell(val points: Set<Pair<Int, Int>>) {
 
     companion object {
         fun fromRoi(roi: Roi): PositionedCell {
             return PositionedCell(roi.containedPoints.map { point ->
-                Pair(point.x.toFloat(), point.y.toFloat())
+                Pair(point.x, point.y)
             }.toSet())
         }
     }
